Resolve `$crate`s for pretty-printing at more appropriate time
